// resources/js/api/generated/models/getApiV1ServersServeridRegistries200DataCredentialsItem.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */

export type GetApiV1ServersServeridRegistries200DataCredentialsItem = {
  created_at: string;
  /** @minimum 0 */
  id: number;
  image_pattern?: string;
  registry_url: string;
  /** @minimum 0 */
  server_id: number;
  stack_pattern: string;
  updated_at: string;
  username: string;
};